ajustes para PR
