+++ fronteend/package.json
@@ -8,7 +8,8 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.14.3",
-    "@mui/material": "^5.14.6",
+    "@mui/material": "^5.14.5",
+    "@mui/x-charts": "^6.0.0-alpha.7",
     "@react-google-maps/api": "^2.19.2",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
