+++ fronteend/src/App.js
@@ -2,8 +2,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { Dashboard } from './pages/Dashboard/';
 import './App.css';
 import PageMap from './pages/Map/map';
-import Signup from './pages/Initial/Signup';
-import Login from './pages/Initial/Login';
+import Login from './pages/Login/Login';
 
 function App() {
   return (
