--- fronteend/src/components/MiniDrawer.jsx
@@ -1,388 +1,0 @@
-import * as React from "react";
-import { useEffect, useState } from "react";
-import { styled, useTheme, alpha } from "@mui/material/styles";
-import { useNavigate } from 'react-router-dom';
-import { Container, Grid } from "@mui/material";
-import axios from "axios";
-import Card from "@mui/material/Card";
-import CardContent from "@mui/material/CardContent";
-import CardMedia from "@mui/material/CardMedia";
-import Box from "@mui/material/Box";
-import MuiDrawer from "@mui/material/Drawer";
-import MuiAppBar from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-import List from "@mui/material/List";
-import CssBaseline from "@mui/material/CssBaseline";
-import Typography from "@mui/material/Typography";
-import InputBase from "@mui/material/InputBase";
-import Divider from "@mui/material/Divider";
-import Badge from "@mui/material/Badge";
-import SearchIcon from "@mui/icons-material/Search";
-import IconButton from "@mui/material/IconButton";
-import MenuIcon from "@mui/icons-material/Menu";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import DashboardIcon from "@mui/icons-material/Dashboard";
-import MapIcon from "@mui/icons-material/Map";
-import AccountCircle from "@mui/icons-material/AccountCircle";
-import SettingsIcon from "@mui/icons-material/Settings";
-import LogoutIcon from "@mui/icons-material/Logout";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-
-const Search = styled("div")(({ theme }) => ({
-  position: "relative",
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.25),
-  "&:hover": {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginRight: theme.spacing(2),
-  marginLeft: 0,
-  width: "100%",
-  [theme.breakpoints.up("sm")]: {
-    marginLeft: theme.spacing(3),
-    width: "auto",
-  },
-}));
-
-
-
-
-const SearchIconWrapper = styled("div")(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: "100%",
-  position: "absolute",
-  pointerEvents: "none",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: "inherit",
-  "& .MuiInputBase-input": {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create("width"),
-    width: "100%",
-    [theme.breakpoints.up("md")]: {
-      width: "30ch",
-    },
-  },
-}));
-
-const drawerWidth = 240;
-
-const openedMixin = (theme) => ({
-  width: drawerWidth,
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: "hidden",
-});
-
-const closedMixin = (theme) => ({
-  transition: theme.transitions.create("width", {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: "hidden",
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up("sm")]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-
-const DrawerHeader = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "flex-end",
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-}));
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  background: "#7AC74F",
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(["width", "margin"], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(["width", "margin"], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  width: drawerWidth,
-  flexShrink: 0,
-  whiteSpace: "nowrap",
-  boxSizing: "border-box",
-  ...(open && {
-    ...openedMixin(theme),
-    "& .MuiDrawer-paper": openedMixin(theme),
-  }),
-  ...(!open && {
-    ...closedMixin(theme),
-    "& .MuiDrawer-paper": closedMixin(theme),
-  }),
-}));
-
-export default function MiniDrawer() {
-  const theme = useTheme();
-  const [open, setOpen] = React.useState(false);
-
-  //const [sensorQtd, setSensorQtd] = useState([]);
-  const [status, setStatus] = useState([]);
-
-
-  const navigate = useNavigate();
-  // useEffect(() => {
-  //     async function getSensorQtd() {
-  //       try {
-  //         let response = await axios.get('https://cata-sucata.azure-api.net/preview/status-preview');
-  //         //console.log(response.data.devices_info.number_devices);
-  //         setSensorQtd(response.data.devices_info);
-  //       } catch (error) {
-  //         console.log('Erro ao buscar dados:', error);
-  //       }
-  //     }
-  //     getSensorQtd();
-  // }, []);
-  //console.log(sensorQtd.number_devices);
-
-  useEffect(() => {
-    async function getStatus() {
-      try {
-        let response = await axios.get(
-          "https://cata-sucata.azure-api.net/preview/status-preview"
-        );
-        //console.log(response.data.status_trash);
-        setStatus(response.data.status_trash);
-        //setSensorQtd(response.data.devices_info);
-      } catch (error) {
-        console.log("Erro ao buscar dados:", error);
-      }
-    }
-    getStatus();
-
-    const interval = setInterval(() => {
-      console.log("Requisição!");
-      getStatus();
-    }, 30000);
-
-    return () => clearInterval(interval);
-  }, []);
-  //console.log(status.full);
-
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
-
-  const handleDrawerClose = () => {
-    setOpen(false);
-  };
-
-  //const menuId = 'primary-search-account-menu';
-
-  return (
-    <Box sx={{ display: "flex" }}>
-      <CssBaseline />
-      <AppBar position="fixed" open={open}>
-        <Toolbar>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            edge="start"
-            sx={{
-              marginRight: 5,
-              ...(open && { display: "none" }),
-            }}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" noWrap component="div">
-            Cata-Sucata
-          </Typography>
-          <Search>
-            <SearchIconWrapper>
-              <SearchIcon />
-            </SearchIconWrapper>
-            <StyledInputBase
-              placeholder="Pesquisar…"
-              inputProps={{ "aria-label": "search" }}
-            />
-          </Search>
-          <Box sx={{ display: { xs: "none", md: "flex" } }}>
-            <IconButton
-              size="large"
-              aria-label="mostra 3 notificações"
-              color="inherit"
-            >
-              <Badge badgeContent={3} color="error">
-                <NotificationsIcon />
-              </Badge>
-            </IconButton>
-          </Box>
-        </Toolbar>
-      </AppBar>
-      <Drawer variant="permanent" open={open}>
-        <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === "rtl" ? (
-              <ChevronRightIcon />
-            ) : (
-              <ChevronLeftIcon />
-            )}
-          </IconButton>
-        </DrawerHeader>
-        <Divider />
-        <List>
-          {["Dashboard", "Rotas", "Perfil", "Configurações", "Sair"].map(
-            (text, index) => (
-              <ListItem key={text} disablePadding sx={{ display: "block" }}>
-                <ListItemButton
-                        sx={{
-                        minHeight: 48,
-                        justifyContent: open ? 'initial' : 'center',
-                        px: 2.5,
-                        }}
-                        onClick={() => {
-                            if (index === 1) {   // Check for the "Rotas" button
-                                navigate('/map');  // Use `navigate` directly
-                            }
-                        }}
-                    >
-                  <ListItemIcon
-                    sx={{
-                      minWidth: 0,
-                      mr: open ? 3 : "auto",
-                      justifyContent: "center",
-                    }}
-                  >
-                    {index === 0 ? <DashboardIcon /> : <></>}
-                    {index === 1 ? <MapIcon /> : <></>}
-                    {index === 2 ? <AccountCircle /> : <></>}
-                    {index === 3 ? <SettingsIcon /> : <></>}
-                    {index === 4 ? <LogoutIcon /> : <></>}
-                  </ListItemIcon>
-                  <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
-                </ListItemButton>
-              </ListItem>
-            )
-          )}
-        </List>
-        <Divider />
-      </Drawer>
-      <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
-        <DrawerHeader />
-        <Container maxWidth="lg">
-          <Grid
-            container
-            spacing={{ lg: 6, md: 3 }}
-            columns={{ xs: 3, sm: 8, md: 10 }}
-          >
-            <Grid item xs={3}>
-              <Card sx={{ maxWidth: 305 }}>
-                <CardMedia
-                  sx={{ height: 10 }}
-                  // image="/static/images/cards/contemplative-reptile.jpg"
-                  // title="green iguana"
-                />
-                <CardContent>
-                  <Box
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography gutterBottom variant="h6" component="div">
-                      Lixeiras cheias
-                    </Typography>
-                    <Typography
-                      color="text.secondary"
-                      gutterBottom
-                      variant="h5"
-                    >
-                      <p>{status.full}</p>
-                    </Typography>
-                  </Box>
-                </CardContent>
-              </Card>
-            </Grid>
-            <Grid item xs={3}>
-              <Card sx={{ maxWidth: 305 }}>
-                <CardMedia
-                  sx={{ height: 10 }}
-                  // image="/static/images/cards/contemplative-reptile.jpg"
-                  // title="green iguana"
-                />
-                <CardContent>
-                  <Box
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography gutterBottom variant="h6" component="div">
-                      Lixeiras médias
-                    </Typography>
-                    <Typography
-                      color="text.secondary"
-                      gutterBottom
-                      variant="h5"
-                    >
-                      <p>{status.medium}</p>
-                    </Typography>
-                  </Box>
-                </CardContent>
-              </Card>
-            </Grid>
-            <Grid item xs={3}>
-              <Card sx={{ maxWidth: 305 }}>
-                <CardMedia
-                  sx={{ height: 10 }}
-                  // image="/static/images/cards/contemplative-reptile.jpg"
-                  // title="green iguana"
-                />
-                <CardContent>
-                  <Box
-                    display="flex"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography gutterBottom variant="h6" component="div">
-                      Lixeiras vazias
-                    </Typography>
-                    <Typography
-                      color="text.secondary"
-                      gutterBottom
-                      variant="h5"
-                    >
-                      <p>{status.empty}</p>
-                    </Typography>
-                  </Box>
-                </CardContent>
-              </Card>
-            </Grid>
-          </Grid>
-        </Container>
-      </Box>
-    </Box>
-  );
-}
